Fixed wrong filter and deprecated versions
Removed --dev from composer because it's deprecated;
Removed PHP versions 5.6 and 7 because PHP is backwords compatible and 5.6 is not officially supported anymore by PHP;
Removed support for WP 3.9 because no official support is present either.
Fixed wrong filters in phpunit.xml that caused double output and wrong folders to be added to clover coverage xml.